refactor: extract SettingsEnterpriseFeatureGateCard shared component
- Extract the Enterprise feature gate card (lock icon + Activate button) into a reusable component
- Use it in both SettingsAIUsageTab and SettingsAdminAI
- Show the gate card instead of the table when Enterprise is not active in admin AI usage section
- Skip usage query when Enterprise is not active

Made-with: Cursor
